Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/selectors/attribute-selectors/attribute-case/cssom-expected.txt


PASS [foo="bar"] /* sanity check */ insertRule
PASS [foo="bar"] /* sanity check */ getting CSSRule#cssText
PASS [foo="bar"] /* sanity check */ getting CSSStyleRule#selectorText
PASS [foo="bar"] /* sanity check */ setting CSSStyleRule#selectorText
PASS [foo="bar"] /* sanity check */ insertRule in @media
PASS [foo="bar"] /* sanity check */ getting CSSRule#cssText in @media
PASS [foo="bar"] /* sanity check */ getting CSSStyleRule#selectorText in @media
PASS [foo="bar"] /* sanity check */ setting CSSStyleRule#selectorText in @media
PASS [foo="bar" i] insertRule
PASS [foo="bar" i] getting CSSRule#cssText
PASS [foo="bar" i] getting CSSStyleRule#selectorText
PASS [foo="bar" i] setting CSSStyleRule#selectorText
PASS [foo="bar" i] insertRule in @media
PASS [foo="bar" i] getting CSSRule#cssText in @media
PASS [foo="bar" i] getting CSSStyleRule#selectorText in @media
PASS [foo="bar" i] setting CSSStyleRule#selectorText in @media
PASS [foo="bar" /**/ i] insertRule
PASS [foo="bar" /**/ i] getting CSSRule#cssText
PASS [foo="bar" /**/ i] getting CSSStyleRule#selectorText
PASS [foo="bar" /**/ i] setting CSSStyleRule#selectorText
PASS [foo="bar" /**/ i] insertRule in @media
PASS [foo="bar" /**/ i] getting CSSRule#cssText in @media
PASS [foo="bar" /**/ i] getting CSSStyleRule#selectorText in @media
PASS [foo="bar" /**/ i] setting CSSStyleRule#selectorText in @media
PASS [foo="bar"/**/i] insertRule
PASS [foo="bar"/**/i] getting CSSRule#cssText
PASS [foo="bar"/**/i] getting CSSStyleRule#selectorText
PASS [foo="bar"/**/i] setting CSSStyleRule#selectorText
PASS [foo="bar"/**/i] insertRule in @media
PASS [foo="bar"/**/i] getting CSSRule#cssText in @media
PASS [foo="bar"/**/i] getting CSSStyleRule#selectorText in @media
PASS [foo="bar"/**/i] setting CSSStyleRule#selectorText in @media
PASS [*|foo="bar" i] insertRule
PASS [*|foo="bar" i] getting CSSRule#cssText
PASS [*|foo="bar" i] getting CSSStyleRule#selectorText
PASS [*|foo="bar" i] setting CSSStyleRule#selectorText
PASS [*|foo="bar" i] insertRule in @media
PASS [*|foo="bar" i] getting CSSRule#cssText in @media
PASS [*|foo="bar" i] getting CSSStyleRule#selectorText in @media
PASS [*|foo="bar" i] setting CSSStyleRule#selectorText in @media
FAIL [foo="bar" s] insertRule The string did not match the expected pattern.
FAIL [foo="bar" s] getting CSSRule#cssText The string did not match the expected pattern.
FAIL [foo="bar" s] getting CSSStyleRule#selectorText The string did not match the expected pattern.
FAIL [foo="bar" s] setting CSSStyleRule#selectorText assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar" s] insertRule in @media The string did not match the expected pattern.
FAIL [foo="bar" s] getting CSSRule#cssText in @media The string did not match the expected pattern.
FAIL [foo="bar" s] getting CSSStyleRule#selectorText in @media The string did not match the expected pattern.
FAIL [foo="bar" s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar" /**/ s] insertRule The string did not match the expected pattern.
FAIL [foo="bar" /**/ s] getting CSSRule#cssText The string did not match the expected pattern.
FAIL [foo="bar" /**/ s] getting CSSStyleRule#selectorText The string did not match the expected pattern.
FAIL [foo="bar" /**/ s] setting CSSStyleRule#selectorText assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar" /**/ s] insertRule in @media The string did not match the expected pattern.
FAIL [foo="bar" /**/ s] getting CSSRule#cssText in @media The string did not match the expected pattern.
FAIL [foo="bar" /**/ s] getting CSSStyleRule#selectorText in @media The string did not match the expected pattern.
FAIL [foo="bar" /**/ s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar"/**/s] insertRule The string did not match the expected pattern.
FAIL [foo="bar"/**/s] getting CSSRule#cssText The string did not match the expected pattern.
FAIL [foo="bar"/**/s] getting CSSStyleRule#selectorText The string did not match the expected pattern.
FAIL [foo="bar"/**/s] setting CSSStyleRule#selectorText assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar"/**/s] insertRule in @media The string did not match the expected pattern.
FAIL [foo="bar"/**/s] getting CSSRule#cssText in @media The string did not match the expected pattern.
FAIL [foo="bar"/**/s] getting CSSStyleRule#selectorText in @media The string did not match the expected pattern.
FAIL [foo="bar"/**/s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [*|foo="bar" s] insertRule The string did not match the expected pattern.
FAIL [*|foo="bar" s] getting CSSRule#cssText The string did not match the expected pattern.
FAIL [*|foo="bar" s] getting CSSStyleRule#selectorText The string did not match the expected pattern.
FAIL [*|foo="bar" s] setting CSSStyleRule#selectorText assert_equals: expected "[*|foo=\"bar\" s]" but got "foobar"
FAIL [*|foo="bar" s] insertRule in @media The string did not match the expected pattern.
FAIL [*|foo="bar" s] getting CSSRule#cssText in @media The string did not match the expected pattern.
FAIL [*|foo="bar" s] getting CSSStyleRule#selectorText in @media The string did not match the expected pattern.
FAIL [*|foo="bar" s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[*|foo=\"bar\" s]" but got "foobar"
